test_code_v2/python/requirements.txt
annotated-types==0.6.0
backoff==2.2.1
britive==2.20.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
cryptography==41.0.3
gql==3.4.0
graphql-core==3.2.3
idna==3.4
install==1.3.5
jmespath==1.0.1
merge-args==0.1.5
multidict==6.0.4
ndjson==0.3.1
pybritive==1.4.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple-typed==3.11.0
python-on-whales==0.70.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==0.10.1
six==1.16.0
tabulate==0.9.0
toml==0.10.2
tqdm==4.66.2
typer==0.10.0
typing_extensions==4.10.0
urllib3==1.26.6
yarl==1.8.2
